+++ rust/shekyl-chain-store/src/store/view.rs
@@ -165,6 +165,8 @@ impl<'id> ChainView<'id> for BatchView<'_, 'id> {
                 hash: info.hash,
                 header: block.header,
                 cumulative_difficulty: info.cumulative_difficulty,
+                coins_generated: info.coins_generated,
+                cumulative_tx_count: info.cumulative_tx_count,
             }),
         })
     }